Rename package ui -> control
