+++ project/app/src/main/AndroidManifest.xml
@@ -10,18 +10,18 @@
         android:label="@string/app_name"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".ui.MainActivity">
+        <activity android:name=".control.MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
                 <category android:name="android.intent.category.LAUNCHER" />
             </intent-filter>
         </activity>
-        <activity android:name=".ui.NewPostActivity"
+        <activity android:name=".control.NewPostActivity"
             android:label="@string/app_name"
-            android:parentActivityName=".ui.MainActivity">
+            android:parentActivityName=".control.MainActivity">
             <meta-data
                 android:name="android.support.PARENT_ACTIVITY"
-                android:value=".ui.MainActivity" />
+                android:value=".control.MainActivity" />
         </activity>
     </application>
