+++ project/app/src/main/java/com/chalmers/tda367/localfeud/control/FeedPagerAdapter.java
@@ -1,4 +1,4 @@
-package com.chalmers.tda367.localfeud.ui;
+package com.chalmers.tda367.localfeud.control;
 
 import android.support.v4.app.Fragment;
 import android.support.v4.app.FragmentManager;
