+++ project/app/src/main/java/com/chalmers/tda367/localfeud/control/PostClickedAdapter.java
@@ -1,4 +1,4 @@
-package com.chalmers.tda367.localfeud.ui;
+package com.chalmers.tda367.localfeud.control;
 
 import android.content.Context;
 import android.support.v7.widget.CardView;
