+++ project/app/src/main/java/com/chalmers/tda367/localfeud/net/ServerComm.java
@@ -6,19 +6,15 @@ import com.chalmers.tda367.localfeud.data.Chat;
 import com.chalmers.tda367.localfeud.data.Position;
 import com.chalmers.tda367.localfeud.data.Post;
 import com.chalmers.tda367.localfeud.data.User;
-import com.chalmers.tda367.localfeud.ui.PostAdapter;
+import com.chalmers.tda367.localfeud.control.PostAdapter;
 import com.chalmers.tda367.localfeud.util.GsonHandler;
 import com.chalmers.tda367.localfeud.util.RestClient;
 import com.chalmers.tda367.localfeud.util.TagHandler;
 import com.chalmers.tda367.localfeud.util.responseActions.IResponseAction;
-import com.loopj.android.http.AsyncHttpResponseHandler;
 
 import java.util.ArrayList;
 import java.util.HashMap;
 import java.util.List;
-
-import cz.msebera.android.httpclient.Header;
-import cz.msebera.android.httpclient.client.HttpResponseException;
 
 /**
  * Created by Alfred on 2016-04-12.
