+++ project/app/src/main/res/layout/activity_main.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.MainActivity">
+    tools:context=".control.MainActivity">
 
     <android.support.design.widget.AppBarLayout
         android:id="@+id/toolbar_layout"
