+++ project/app/src/main/res/layout/activity_post_clicked.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.PostClickedActivity">
+    tools:context=".control.PostClickedActivity">
 
     <android.support.v7.widget.RecyclerView
         android:id="@+id/comment_feed_recyclerview"
